% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeJags.R
\name{replaceScientificNotationR}
\alias{replaceScientificNotationR}
\title{Internal Function to write JAGS model}
\description{
This is an internally called function used to write
  theJAGS model. Modified from Andrew Gelman function.
}
\keyword{internal}
