% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeJags.R
\name{writeJags}
\alias{writeJags}
\title{BUGS to JAGS}
\usage{
writeJags(model, filename, dir, overwrite, digits)
}
\arguments{
\item{model}{JAGS model format to be write onto the hard drive.}

\item{filename}{character,  the name of the file to write/overwrite.}

\item{dir}{directory where to write the file. Defaults to uses the
current working directory through \code{getwd}}

\item{overwrite}{logical, if \code{TRUE} the \code{filename} with
the same name will be overwrited.}

\item{digits}{Number of digits used in the output.}
}
\description{
Writes a BUGS model file in a JAGS format
}
\examples{
\dontrun{
 AnimalModelBUGSVar <- function()
  {
   ## Priors - precision parameters
   tau2e ~ dgamma(0.001, 0.001)
   tau2a ~ dgamma(0.001, 0.001)
   sigma2e <- pow(tau2e, -1)
   sigma2a <- pow(tau2a, -1)
   ## Additive genetic values
   for(k in 1:nI) {
    a[id[k]] ~ dnorm(pa[id[k]], Xtau2a[id[k]])
    pa[id[k]] <- 0.5 * (a[fid[k]] + a[mid[k]])
    Xtau2a[id[k]] <- winv[id[k]] * tau2a
   }
   a[nU] <- 0 # NULL (zero) holder
   ## Fixed Effects
   for(j in 1:nB) { b[j] ~ dnorm(0, 1.0E-6) }
    ## Phenotypes
    for(i in 1:nY) {
     y[i] ~ dnorm(mu[i], tau2e)
     mu[i] <- inprod(x[i, ], b[]) + a[idy[i]]
    }
  }

filename <- "model.txt"

writeJags(model = AnimalModelBUGSVar, filename = filename)
}

}
\seealso{
\code{\link{AnimalModelBUGSData}}
}
\author{
Thiago de Paula Oliveira, \email{thiago.oliveira@ed.ac.uk},
  Ivan Pocrnic, \email{ivan.pocrnic@roslin.ed.ac.uk}, Gregor
  Gorjanc, \email{gregor.gorjanc@gmail.com}
}
\keyword{BUGS}
\keyword{JAGS}
